+++ src/components/model.go
@@ -2,32 +2,31 @@ package components
 
 import (
 	"encoding/json"
-	"log"
-	"os"
-	"path/filepath"
-
 	"github.com/barasher/go-exiftool"
 	"github.com/charmbracelet/bubbles/textinput"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
 	"github.com/rkoesters/xdg/basedir"
+	"log"
+	"os"
+	"path/filepath"
 )
 
 const (
-	configFolder     string = "/config"
 	themeFolder      string = "/theme"
 	dataFolder       string = "/data"
-	lastCheckVersion string = "/data/lastCheckVersion"
-	pinnedFile       string = "/data/pinned.json"
+	lastCheckVersion string = "/lastCheckVersion"
+	pinnedFile       string = "/pinned.json"
+	toggleDotFile    string = "/toggleDotFile"
 	logFile          string = "/superfile.log"
-	configFile       string = "/config/config.json"
+	configFile       string = "/config.json"
 	themeZipName     string = "/theme.zip"
 )
 
 var HomeDir = basedir.Home
 var SuperFileMainDir = basedir.ConfigHome + "/superfile"
 var SuperFileCacheDir = basedir.CacheHome + "/superfile"
-
+var SuperFileDataDir = basedir.DataHome + "/superfile"
 var theme ThemeType
 var Config ConfigType
 
@@ -38,7 +37,8 @@ var channel = make(chan channelMessage, 1000)
 
 func InitialModel(dir string) model {
 	var err error
-	logOutput, err = os.OpenFile(SuperFileCacheDir + logFile, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+	logOutput, err = os.OpenFile(SuperFileCacheDir+logFile, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+
 	if err != nil {
 		log.Fatalf("Error while opening superfile.log file: %v", err)
 	}
@@ -51,7 +51,7 @@ func InitialModel(dir string) model {
 	err = json.Unmarshal(data, &Config)
 
 	if err != nil {
-		log.Fatalf("Error decoding config json( your config file may have misconfigured ): %v", err)
+		log.Fatalf("Error decoding config json (your config file may have misconfigured): %v", err)
 	}
 
 	data, err = os.ReadFile(SuperFileMainDir + themeFolder + "/" + Config.Theme + ".json")
@@ -64,7 +64,11 @@ func InitialModel(dir string) model {
 	if err != nil {
 		log.Fatalf("Error while decoding theme json( Your theme file may have errors ): %v", err)
 	}
-
+	toggleDotFileData, err := os.ReadFile(SuperFileDataDir + toggleDotFile)
+	if err != nil {
+		OutPutLog("Error while reading toggleDotFile data error:", err)
+	}
+	var toggleDotFileBool = string(toggleDotFileData) == "true"
 	LoadThemeConfig()
 	et, err = exiftool.NewExiftool()
 	if err != nil {
@@ -103,6 +107,7 @@ func InitialModel(dir string) model {
 			},
 			width: 10,
 		},
+		toggleDotFile: toggleDotFileBool,
 	}
 }
 
@@ -247,6 +252,8 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 				m = PinnedFolder(m)
 			case Config.OpenTerminal[0], Config.OpenTerminal[1]:
 				m = OpenTerminal(m)
+			case Config.ToggleDotFile[0], Config.ToggleDotFile[1]:
+				m = ToggleDotFile(m)
 			default:
 				// check if it's the select mode
 				if m.fileModel.filePanels[m.filePanelFocusIndex].focusType == focus && m.fileModel.filePanels[m.filePanelFocusIndex].panelMode == selectMode {
@@ -318,7 +325,7 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 }
 
 func (m model) View() string {
-	// check is the terminal size enough
+	// Check if terminal dimensions are big enough
 	if m.fullHeight < minimumHeight || m.fullWidth < minimumWidth {
 		return TerminalSizeWarnRender(m)
 	} else if m.typingModal.open {
@@ -327,20 +334,12 @@ func (m model) View() string {
 		return WarnModalRender(m)
 	} else {
 		sideBar := SideBarRender(m)
-
 		filePanel := FilePanelRender(m)
-
 		mainPanel := lipgloss.JoinHorizontal(0, sideBar, filePanel)
-
 		processBar := ProcessBarRender(m)
-
 		metaData := MetaDataRender(m)
-
 		clipboardBar := ClipboardRender(m)
-
 		bottomBar := lipgloss.JoinHorizontal(0, processBar, metaData, clipboardBar)
-
-		// final render
 		finalRender := lipgloss.JoinVertical(0, mainPanel, bottomBar)
 
 		return lipgloss.JoinVertical(lipgloss.Top, finalRender)